api/requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask_RESTful==0.3.10
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.0
requests==2.32.3
urllib3==2.2.2
Werkzeug==3.0.3
flask_pymongo==2.3.0
python-dotenv==1.0.0
mongoengine==0.28.2
sfapi_client==0.1.0
Authlib==1.3.2
matplotlib==3.9.2
numpy==2.0.0
ase==3.23.0
scipy==1.14.1
